requirements.txt
numpy>=1.19.0
pandas>=1.1.0
python-Levenshtein>=0.20.8
requests>=2.25.0 